P5-5: Decompose SessionPane.tsx (1009→269 LOC) — extract 6 components + hook
Extracted:
- SessionMount.tsx (152 LOC) — xterm terminal mount for one PTY session
- AIChatTab.tsx (168 LOC) — pinned AI Chat tab + thread history dropdown
- AgentSessionTab.tsx (147 LOC) — pinned Cairn Agent tab + session dropdown
- TerminalTab.tsx (75 LOC) — single PTY session tab button
- AgentEmptyState.tsx (70 LOC) — no-session empty state with recent list
- useAgentSessionActions.ts (96 LOC) — shared create/resume session hook
- sessionUtils.ts (10 LOC) — formatDate helper

SessionPane.tsx is now a 269-LOC orchestrator that composes tab bar +
content switching + top-level effects.

type-check:all, lint, 430 tests pass.
