+++ src/components/agent/sessionUtils.ts
@@ -1,0 +1,10 @@
+// Pure helper for formatting session timestamps.
+
+export function formatDate(iso: string) {
+  const d = new Date(iso);
+  const diffDays = Math.floor((Date.now() - d.getTime()) / 86400000);
+  if (diffDays === 0) return "Today";
+  if (diffDays === 1) return "Yesterday";
+  if (diffDays < 7) return `${diffDays}d ago`;
+  return d.toLocaleDateString(undefined, { month: "short", day: "numeric" });
+}
